Split ContractsBCL and move it to shared.
