+++ src/System.Private.CoreLib/shared/System/Runtime/CompilerServices/ContractHelper.cs
@@ -1,0 +1,190 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+#define DEBUG // The behavior of this contract library should be consistent regardless of build type.
+
+using System;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.Diagnostics.CodeAnalysis;
+using System.Diagnostics.Contracts;
+using System.Reflection;
+
+namespace System.Runtime.CompilerServices
+{
+    public static partial class ContractHelper
+    {
+#region Private fields
+
+        private static volatile EventHandler<ContractFailedEventArgs> contractFailedEvent;
+        private static readonly object lockObject = new object();
+
+        internal const int COR_E_CODECONTRACTFAILED = unchecked((int)0x80131542);
+
+#endregion
+
+        /// <summary>
+        /// Allows a managed application environment such as an interactive interpreter (IronPython) or a
+        /// web browser host (Jolt hosting Silverlight in IE) to be notified of contract failures and 
+        /// potentially "handle" them, either by throwing a particular exception type, etc.  If any of the
+        /// event handlers sets the Cancel flag in the ContractFailedEventArgs, then the Contract class will
+        /// not pop up an assert dialog box or trigger escalation policy.  Hooking this event requires 
+        /// full trust.
+        /// </summary>
+        internal static event EventHandler<ContractFailedEventArgs> InternalContractFailed
+        {
+            add
+            {
+                // Eagerly prepare each event handler _marked with a reliability contract_, to 
+                // attempt to reduce out of memory exceptions while reporting contract violations.
+                // This only works if the new handler obeys the constraints placed on 
+                // constrained execution regions.  Eagerly preparing non-reliable event handlers
+                // would be a perf hit and wouldn't significantly improve reliability.
+                // UE: Please mention reliable event handlers should also be marked with the 
+                // PrePrepareMethodAttribute to avoid CER eager preparation work when ngen'ed.
+                System.Runtime.CompilerServices.RuntimeHelpers.PrepareContractedDelegate(value);
+                lock (lockObject)
+                {
+                    contractFailedEvent += value;
+                }
+            }
+            remove
+            {
+                lock (lockObject)
+                {
+                    contractFailedEvent -= value;
+                }
+            }
+        }
+
+        /// <summary>
+        /// Rewriter will call this method on a contract failure to allow listeners to be notified.
+        /// The method should not perform any failure (assert/throw) itself.
+        /// This method has 3 functions:
+        /// 1. Call any contract hooks (such as listeners to Contract failed events)
+        /// 2. Determine if the listeners deem the failure as handled (then resultFailureMessage should be set to null)
+        /// 3. Produce a localized resultFailureMessage used in advertising the failure subsequently.
+        /// On exit: null if the event was handled and should not trigger a failure.
+        ///          Otherwise, returns the localized failure message.
+        /// </summary>
+        [System.Diagnostics.DebuggerNonUserCode]
+        public static string RaiseContractFailedEvent(ContractFailureKind failureKind, string userMessage, string conditionText, Exception innerException)
+        {
+            if (failureKind < ContractFailureKind.Precondition || failureKind > ContractFailureKind.Assume)
+                throw new ArgumentException(SR.Format(SR.Arg_EnumIllegalVal, failureKind), nameof(failureKind));
+
+            string returnValue;
+            string displayMessage = "contract failed.";  // Incomplete, but in case of OOM during resource lookup...
+            ContractFailedEventArgs eventArgs = null;  // In case of OOM.
+
+            try
+            {
+                displayMessage = GetDisplayMessage(failureKind, userMessage, conditionText);
+                EventHandler<ContractFailedEventArgs> contractFailedEventLocal = contractFailedEvent;
+                if (contractFailedEventLocal != null)
+                {
+                    eventArgs = new ContractFailedEventArgs(failureKind, displayMessage, conditionText, innerException);
+                    foreach (EventHandler<ContractFailedEventArgs> handler in contractFailedEventLocal.GetInvocationList())
+                    {
+                        try
+                        {
+                            handler(null, eventArgs);
+                        }
+                        catch (Exception e)
+                        {
+                            eventArgs.thrownDuringHandler = e;
+                            eventArgs.SetUnwind();
+                        }
+                    }
+                    if (eventArgs.Unwind)
+                    {
+                        // unwind
+                        if (innerException == null) { innerException = eventArgs.thrownDuringHandler; }
+                        throw new ContractException(failureKind, displayMessage, userMessage, conditionText, innerException);
+                    }
+                }
+            }
+            finally
+            {
+                if (eventArgs != null && eventArgs.Handled)
+                {
+                    returnValue = null; // handled
+                }
+                else
+                {
+                    returnValue = displayMessage;
+                }
+            }
+            return returnValue;
+        }
+
+        private static string GetResourceNameForFailure(ContractFailureKind failureKind)
+        {
+            string resourceName = null;
+            switch (failureKind)
+            {
+                case ContractFailureKind.Assert:
+                    resourceName = "AssertionFailed";
+                    break;
+
+                case ContractFailureKind.Assume:
+                    resourceName = "AssumptionFailed";
+                    break;
+
+                case ContractFailureKind.Precondition:
+                    resourceName = "PreconditionFailed";
+                    break;
+
+                case ContractFailureKind.Postcondition:
+                    resourceName = "PostconditionFailed";
+                    break;
+
+                case ContractFailureKind.Invariant:
+                    resourceName = "InvariantFailed";
+                    break;
+
+                case ContractFailureKind.PostconditionOnException:
+                    resourceName = "PostconditionOnExceptionFailed";
+                    break;
+
+                default:
+                    Debug.Fail("Unreachable code");
+                    resourceName = "AssumptionFailed";
+                    break;
+            }
+            return resourceName;
+        }
+
+        private static string GetDisplayMessage(ContractFailureKind failureKind, string userMessage, string conditionText)
+        {
+            string failureMessage;
+            // Well-formatted English messages will take one of four forms.  A sentence ending in
+            // either a period or a colon, the condition string, then the message tacked 
+            // on to the end with two spaces in front.
+            // Note that both the conditionText and userMessage may be null.  Also, 
+            // on Silverlight we may not be able to look up a friendly string for the
+            // error message.  Let's leverage Silverlight's default error message there. 
+            if (!string.IsNullOrEmpty(conditionText))
+            {
+                string resourceName = GetResourceNameForFailure(failureKind); 
+                resourceName += "_Cnd";
+                failureMessage = SR.Format(SR.GetResourceString(resourceName), conditionText);
+            }
+            else
+            {
+                failureMessage = "";
+            }
+
+            // Now add in the user message, if present.
+            if (!string.IsNullOrEmpty(userMessage))
+            {
+                return failureMessage + "  " + userMessage;
+            }
+            else
+            {
+                return failureMessage;
+            }
+        }
+    }
+}
