+++ src/System.Private.CoreLib/src/System.Private.CoreLib.csproj
@@ -197,7 +197,7 @@
     <Compile Include="System\Collections\Generic\EqualityComparer.CoreRT.cs" />
     <Compile Include="System\Collections\Generic\EqualOnlyComparer.cs" />
     <Compile Include="System\DefaultBinder.CanConvert.cs" />
-    <Compile Include="System\Diagnostics\Contracts\ContractsBCL.cs" />
+    <Compile Include="System\Runtime\CompilerServices\ContractHelper.CoreRT.cs" />
     <Compile Include="System\Globalization\GlobalizationMode.cs" />
     <Compile Include="System\Globalization\GlobalizationMode.Windows.cs" Condition="'$(TargetsWindows)'=='true'" />
     <Compile Include="System\Globalization\GlobalizationMode.Unix.cs" Condition="'$(TargetsUnix)'=='true'" />
@@ -333,8 +333,8 @@
     <Compile Condition="'$(EnableWinRT)' == 'true'" Include="System\TimeZoneInfo.WinRT.cs" />
     <Compile Include="System\AppContext.Windows.CoreRT.cs" />
     <Compile Include="System\DateTime.Windows.CoreRT.cs" />
-    <Compile Include="System\DateTime.Win32.cs" Condition="'$(EnableWinRT)'!='true'"/>
-    <Compile Include="System\DateTime.WinRT.cs" Condition="'$(EnableWinRT)'=='true'"/>
+    <Compile Include="System\DateTime.Win32.cs" Condition="'$(EnableWinRT)'!='true'" />
+    <Compile Include="System\DateTime.WinRT.cs" Condition="'$(EnableWinRT)'=='true'" />
     <Compile Include="System\Runtime\InteropServices\PInvokeMarshal.Windows.cs" />
     <Compile Include="System\Runtime\MemoryFailPoint.cs" />
     <Compile Include="..\..\Common\src\Interop\Windows\kernel32\Interop.FileTimeToSystemTime.cs">
@@ -673,6 +673,9 @@
     <EmbeddedResource Include="Resources\$(AssemblyName).rd.xml">
       <LogicalName>Resources.$(AssemblyName).rd.xml</LogicalName>
     </EmbeddedResource>
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="System\Diagnostics\Contracts\" />
   </ItemGroup>
   <Import Project="..\shared\System.Private.CoreLib.Shared.projitems" Label="Shared" />
   <Import Project="$([MSBuild]::GetDirectoryNameOfFileAbove($(MSBuildThisFileDirectory), dir.targets))\dir.targets" />
